Added print in configure for debugging

+++ install/scripts/build.sh
@@ -5,6 +5,11 @@ BUILD=_build/${TARGET}
 rm -rf -- ${BUILD}
 mkdir ${BUILD} || exit 1
 tar -zxf Downloads/${TARGET}.tar.gz --strip-components=1 --directory=${BUILD} || exit 1
-_install || exit 1
+_install
+if [[ $? -ne 0 ]]
+then
+  cat _build/${TARGET}.log
+  exit 1
+fi
 rm -rf -- ${BUILD} _build/${TARGET}.log
 exit 0
